feat!: Add `no_std` support (#57)
The `GetSize` trait itself only needs `core::mem::size_of`, so the crate is now
`#![no_std]` with the standard library implementations behind features.

- `std` (default) implies `alloc`, both gate the respective implementations
- the implementations are regrouped into `core_impls`, `alloc_impls` and
  `std_impls`, so a `no_std` build drops whole modules instead of threading
  `cfg` attributes through every file
- `StandardTracker` needs an allocation and is therefore gated behind `alloc`,
  falling back to a `BTreeSet` when `std` is unavailable
- the derive macro emits `::get_size2::default_tracker()`, which resolves to
  `StandardTracker` or, without `alloc`, to `NoTracker`
- atomics are gated per width via `target_has_atomic`, extending the existing
  `AtomicI64`/`AtomicU64` gating to bare metal targets lacking the smaller ones
- `dashmap` and `parking_lot` require `std` and enable it implicitly, the new
  `all-features-no-std` pack activates everything which works without `std`
- CI builds `thumbv7m-none-eabi`, core only and with every `no_std` feature

Two bugs found along the way are fixed as well:

- a tuple struct field handled by `ignore`, `size` or `size_fn` did not advance
  the positional counter, so every following field was read at the wrong
  position: `struct S(#[get_size(ignore)] String, String)` measured the ignored
  field instead of the second one. Deriving both cases from a single accessor
  also lets `size_fn` work on unnamed fields.
- `&CStr` and `&OsStr` counted the bytes they only borrow, which contradicts the
  ownership based accounting and was inconsistent with `&str`, `&Path`, `&[T]`
  and `&T`, all of which report zero.

BREAKING CHANGE: `default = ["std"]` instead of `default = []`. Dependents using
`default-features = false` previously still got every implementation and now have
to enable the `std` feature explicitly. `<&CStr>::get_heap_size()` and
`<&OsStr>::get_heap_size()` now return 0; measure the owned `CString`/`OsString`
to account for those bytes.
