+++ crates/get-size-derive2/src/lib.rs
@@ -239,7 +239,7 @@ fn derive_get_size_impl(input: TokenStream) -> syn::Result<TokenStream> {
             let generated = quote! {
                 impl #impl_generics ::get_size2::GetSize for #name #ty_generics #where_clause {
                     fn get_heap_size(&self) -> usize {
-                        let tracker = get_size2::StandardTracker::default();
+                        let tracker = ::get_size2::default_tracker();
 
                         let (total, _) = ::get_size2::GetSize::get_heap_size_with_tracker(self, tracker);
 
@@ -280,7 +280,19 @@ fn derive_get_size_impl(input: TokenStream) -> syn::Result<TokenStream> {
                 let attr = StructFieldAttribute::from_attributes(&field.attrs)
                     .map_err(|err| syn::Error::new_spanned(field, err.to_string()))?;
 
-                // NOTE There will be no attributes if this is a tuple struct.
+                // How this field is accessed: by name, or by position for tuple structs. The
+                // positional counter has to advance even when the field is handled by one of the
+                // attributes below, otherwise all following fields of a tuple struct would be
+                // read at the wrong position.
+                let accessor = field.ident.as_ref().map_or_else(
+                    || {
+                        let index = syn::Index::from(unidentified_fields_count);
+                        unidentified_fields_count += 1;
+                        quote! { #index }
+                    },
+                    |ident| quote! { #ident },
+                );
+
                 if let Some(size) = attr.size {
                     cmds.push(quote! {
                         total += #size;
@@ -288,15 +300,8 @@ fn derive_get_size_impl(input: TokenStream) -> syn::Result<TokenStream> {
 
                     continue;
                 } else if let Some(size_fn) = attr.size_fn {
-                    let ident = field.ident.as_ref().ok_or_else(|| {
-                        syn::Error::new_spanned(
-                            field,
-                            "get_size(size_fn = ...) is only supported on named fields",
-                        )
-                    })?;
-
                     cmds.push(quote! {
-                        total += #size_fn(&self.#ident);
+                        total += #size_fn(&self.#accessor);
                     });
 
                     continue;
@@ -304,27 +309,17 @@ fn derive_get_size_impl(input: TokenStream) -> syn::Result<TokenStream> {
                     continue;
                 }
 
-                if let Some(ident) = field.ident.as_ref() {
-                    cmds.push(quote! {
-                        let (total_add, tracker) = ::get_size2::GetSize::get_heap_size_with_tracker(&self.#ident, tracker);
-                        total += total_add;
-                    });
-                } else {
-                    let current_index = syn::Index::from(unidentified_fields_count);
-                    cmds.push(quote! {
-                        let (total_add, tracker) = ::get_size2::GetSize::get_heap_size_with_tracker(&self.#current_index, tracker);
-                        total += total_add;
-                    });
-
-                    unidentified_fields_count += 1;
-                }
+                cmds.push(quote! {
+                    let (total_add, tracker) = ::get_size2::GetSize::get_heap_size_with_tracker(&self.#accessor, tracker);
+                    total += total_add;
+                });
             }
 
             // Build the trait implementation
             let generated = quote! {
                 impl #impl_generics ::get_size2::GetSize for #name #ty_generics #where_clause {
                     fn get_heap_size(&self) -> usize {
-                        let tracker = get_size2::StandardTracker::default();
+                        let tracker = ::get_size2::default_tracker();
 
                         let (total, _) = ::get_size2::GetSize::get_heap_size_with_tracker(self, tracker);
