+++ crates/get-size2/src/impls/core_impls/mod.rs
@@ -1,0 +1,5 @@
+mod primitives;
+mod ranges;
+mod references;
+mod tuples;
+mod wrappers;
