+++ crates/get-size2/src/impls/core_impls/primitives.rs
@@ -1,18 +1,24 @@
-use std::convert::Infallible;
-use std::marker::{PhantomData, PhantomPinned};
-use std::num::{
+use core::convert::Infallible;
+use core::marker::{PhantomData, PhantomPinned};
+use core::num::{
     NonZeroI8, NonZeroI16, NonZeroI32, NonZeroI64, NonZeroI128, NonZeroIsize, NonZeroU8,
     NonZeroU16, NonZeroU32, NonZeroU64, NonZeroU128, NonZeroUsize,
 };
-use std::sync::atomic::{
-    AtomicBool, AtomicI8, AtomicI16, AtomicI32, AtomicIsize, AtomicU8, AtomicU16, AtomicU32,
-    AtomicUsize, Ordering,
-};
-// `AtomicI64`/`AtomicU64` are only available on targets that support 64-bit atomics.
-// 32-bit targets (e.g. ppc32) lack them, so they must be gated. See issue #54.
+use core::sync::atomic::Ordering;
+// An atomic type only exists on targets whose atomic support matches its width. Bare metal
+// targets in particular may lack some of them, as do 32-bit targets (e.g. ppc32) for the 64-bit
+// atomics. See issue #54.
+#[cfg(target_has_atomic = "8")]
+use core::sync::atomic::{AtomicBool, AtomicI8, AtomicU8};
+#[cfg(target_has_atomic = "16")]
+use core::sync::atomic::{AtomicI16, AtomicU16};
+#[cfg(target_has_atomic = "32")]
+use core::sync::atomic::{AtomicI32, AtomicU32};
 #[cfg(target_has_atomic = "64")]
-use std::sync::atomic::{AtomicI64, AtomicU64};
-use std::time::{Duration, Instant, SystemTime};
+use core::sync::atomic::{AtomicI64, AtomicU64};
+#[cfg(target_has_atomic = "ptr")]
+use core::sync::atomic::{AtomicIsize, AtomicUsize};
+use core::time::Duration;
 
 use crate::GetSize;
 
@@ -46,26 +52,33 @@ impl GetSize for f32 {}
 impl GetSize for f64 {}
 impl GetSize for char {}
 
+#[cfg(target_has_atomic = "8")]
 impl GetSize for AtomicBool {}
+#[cfg(target_has_atomic = "8")]
 impl GetSize for AtomicI8 {}
+#[cfg(target_has_atomic = "16")]
 impl GetSize for AtomicI16 {}
+#[cfg(target_has_atomic = "32")]
 impl GetSize for AtomicI32 {}
 #[cfg(target_has_atomic = "64")]
 impl GetSize for AtomicI64 {}
+#[cfg(target_has_atomic = "ptr")]
 impl GetSize for AtomicIsize {}
+#[cfg(target_has_atomic = "8")]
 impl GetSize for AtomicU8 {}
+#[cfg(target_has_atomic = "16")]
 impl GetSize for AtomicU16 {}
+#[cfg(target_has_atomic = "32")]
 impl GetSize for AtomicU32 {}
 #[cfg(target_has_atomic = "64")]
 impl GetSize for AtomicU64 {}
+#[cfg(target_has_atomic = "ptr")]
 impl GetSize for AtomicUsize {}
 impl GetSize for Ordering {}
-impl GetSize for std::cmp::Ordering {}
+impl GetSize for core::cmp::Ordering {}
 
 impl GetSize for Infallible {}
 impl<T> GetSize for PhantomData<T> {}
 impl GetSize for PhantomPinned {}
 
-impl GetSize for Instant {}
 impl GetSize for Duration {}
-impl GetSize for SystemTime {}
