+++ crates/get-size2/src/impls/feature/hashbrown.rs
@@ -2,9 +2,9 @@ use crate::{GetSize, GetSizeTracker};
 
 impl<K, V, H> GetSize for hashbrown::HashMap<K, V, H>
 where
-    K: GetSize + Eq + std::hash::Hash,
+    K: GetSize + Eq + core::hash::Hash,
     V: GetSize,
-    H: std::hash::BuildHasher,
+    H: core::hash::BuildHasher,
 {
     fn get_heap_size_with_tracker<Tr: GetSizeTracker>(&self, tracker: Tr) -> (usize, Tr) {
         let (size, tracker) = self
@@ -21,8 +21,8 @@ where
 
 impl<T, H> GetSize for hashbrown::HashSet<T, H>
 where
-    T: GetSize + Eq + std::hash::Hash,
-    H: std::hash::BuildHasher,
+    T: GetSize + Eq + core::hash::Hash,
+    H: core::hash::BuildHasher,
 {
     fn get_heap_size_with_tracker<Tr: GetSizeTracker>(&self, tracker: Tr) -> (usize, Tr) {
         let (size, tracker) = self.iter().fold((0, tracker), |(size, tracker), element| {
