+++ crates/get-size2/src/impls/std_impls/mod.rs
@@ -1,0 +1,3 @@
+mod collections;
+mod sync;
+mod types;
