+++ crates/get-size2/src/lib.rs
@@ -1,5 +1,11 @@
 #![doc = include_str!("./lib.md")]
 #![cfg_attr(docsrs, feature(doc_cfg))]
+#![no_std]
+
+#[cfg(feature = "alloc")]
+extern crate alloc;
+#[cfg(feature = "std")]
+extern crate std;
 
 #[cfg(feature = "derive")]
 #[cfg_attr(docsrs, doc(cfg(feature = "derive")))]
@@ -8,7 +14,9 @@ pub use get_size_derive2::*;
 mod impls;
 mod tracker;
 pub use tracker::*;
-#[cfg(test)]
+// The test suite exercises the `std` implementations and is therefore only
+// compiled when the `std` feature is active.
+#[cfg(all(test, feature = "std"))]
 mod tests;
 
 /// Determines how many bytes the object occupies inside the heap.
@@ -16,14 +24,51 @@ pub fn heap_size<T: GetSize>(value: &T) -> usize {
     value.get_heap_size()
 }
 
+/// The tracker used by the [`GetSize`] implementations generated by the derive macro.
+///
+/// This is [`StandardTracker`] if the `alloc` feature is active, since tracking requires an
+/// allocated set of already seen addresses. Without `alloc` there is no shared ownership type
+/// (`Rc`/`Arc`) to track, so [`NoTracker`] is used instead.
+#[cfg(feature = "alloc")]
+#[cfg_attr(docsrs, doc(cfg(feature = "alloc")))]
+pub type DefaultTracker = StandardTracker;
+
+/// The tracker used by the [`GetSize`] implementations generated by the derive macro.
+///
+/// This is [`StandardTracker`] if the `alloc` feature is active, since tracking requires an
+/// allocated set of already seen addresses. Without `alloc` there is no shared ownership type
+/// (`Rc`/`Arc`) to track, so [`NoTracker`] is used instead.
+#[cfg(not(feature = "alloc"))]
+pub type DefaultTracker = NoTracker;
+
+/// Creates the [`DefaultTracker`] used by the derive macro.
+///
+/// Mostly an implementation detail of the derive macro, which cannot know whether the `alloc`
+/// feature is active at the call site.
+#[cfg(feature = "alloc")]
+#[must_use]
+pub fn default_tracker() -> DefaultTracker {
+    StandardTracker::new()
+}
+
+/// Creates the [`DefaultTracker`] used by the derive macro.
+///
+/// Mostly an implementation detail of the derive macro, which cannot know whether the `alloc`
+/// feature is active at the call site.
+#[cfg(not(feature = "alloc"))]
+#[must_use]
+pub const fn default_tracker() -> DefaultTracker {
+    NoTracker::new(true)
+}
+
 /// Determine the size in bytes an object occupies inside RAM.
 pub trait GetSize: Sized {
     /// Determines how may bytes this object occupies inside the stack.
     ///
-    /// The default implementation uses [`std::mem::size_of`] and should work for almost all types.
+    /// The default implementation uses [`core::mem::size_of`] and should work for almost all types.
     #[must_use]
     fn get_stack_size() -> usize {
-        std::mem::size_of::<Self>()
+        core::mem::size_of::<Self>()
     }
 
     /// Determines how many bytes this object occupies inside the heap.
